Re-add WIN32 ifdef to prevent unknown pragma warnings in gcc.
Looks like this warning pragma is a Visual C++ feature and produces
unknown pragma warnings in gcc. So I reverted this change back to
the original ifdef. Maybe the pragma isn't needed at all for current
Visual C++ compiler, then it should be removed. Or maybe the code
which produces the warning in the first place should be fixed instead.